[UPD] Classes Make, com a adição de uma classe básica para todos os modelos
